+++ app/Http/Controllers/Api/V1/UserController.php
@@ -31,14 +31,11 @@ class UserController extends ApiController
      */
     public function store(StoreUserRequest $request)
     {
-        try {
-            // policy
-            $this->isAble('store', User::class);
-
+        if ($this->isAble('store', User::class)) {
             return new UserResource(User::create($request->mappedAttributes()));
-        } catch (AuthorizationException $ex) {
-            return $this->errorResponse('You are not authorized to create that resource', 401);
         }
+
+        return $this->errorResponse('You are not authorized to create that resource', 401);
     }
 
     /**
@@ -62,15 +59,15 @@ class UserController extends ApiController
             $user = User::findOrFail($user_id);
 
             // policy
-            $this->isAble('update', $user);
+            if ($this->isAble('update', $user)) {
+                $user->update($request->mappedAttributes());
 
-            $user->update($request->mappedAttributes());
+                return new UserResource($user);
+            }
 
-            return new UserResource($user);
-        } catch (ModelNotFoundException $exception) {
-            return $this->errorResponse('Ticket cannot be found.', 404);
-        } catch (AuthorizationException $ex) {
             return $this->errorResponse('You are not authorized to update that resource', 401);
+        } catch (ModelNotFoundException $exception) {
+            return $this->errorResponse('User cannot be found.', 404);
         }
     }
 
@@ -80,12 +77,12 @@ class UserController extends ApiController
             $user = User::findOrFail($user_id);
 
             // policy
-            $this->isAble('replace', $user);
+            if ($this->isAble('replace', $user)) {
+                $user->update($request->mappedAttributes());
 
-            $user->update($request->mappedAttributes());
-
-
-            return new UserResource($user);
+                return new UserResource($user);
+            }
+            return $this->errorResponse('You are not authorized to update that resource', 401);
         } catch (ModelNotFoundException $exception) {
             return $this->errorResponse('User cannot be found.', 404);
         }
@@ -100,11 +97,13 @@ class UserController extends ApiController
             $user = User::findOrFail($user_id);
 
             // policy
-            $this->isAble('delete', $user);
+            if ($this->isAble('delete', $user)) {
+                $user->delete();
 
-            $user->delete();
+                return $this->ok('User successfully deleted');
+            }
+            return $this->errorResponse('You are not authorized to delete that resource', 401);
 
-            return $this->ok('User successfully deleted');
         } catch (ModelNotFoundException $exception) {
             return $this->errorResponse('User cannot found.', 404);
         }
